Remove RoleRouter and update user role checks in UserRouter; add user deletion endpoint
