+++ main.py
@@ -7,7 +7,6 @@ from routers.v1.BookRouter import BookRouter
 from models import UserModel
 from routers.v1.UserRouter import AuthRouter
 
-from routers.v1.RoleRouter import RoleRouter
 
 
 app = FastAPI()
@@ -17,4 +16,3 @@ UserModel.Base.metadata.create_all(bind=engine)
 
 app.include_router(AuthRouter, prefix="/v1/auth", tags=["auth"])
 app.include_router(BookRouter, prefix="/v1/book", tags=["book"])
-app.include_router(RoleRouter, prefix="/v1/role", tags=["role"])
